+++ src/io/exports/dxfExport.cpp
@@ -70,7 +70,7 @@ bool dxfExport::end()
 	return m_status;
 }
 
-bool dxfExport::text(std::wstring name, int layer, const glm::vec3& center, float radius)
+bool dxfExport::text(std::wstring name, int layer, const glm::dvec3& center, double radius)
 {
 	m_file << "  " << 0 << std::endl; // entity type
 	m_file << "TEXT" << std::endl;
@@ -97,7 +97,7 @@ bool dxfExport::text(std::wstring name, int layer, const glm::vec3& center, floa
 	return (m_file.bad());
 }
 
-bool dxfExport::point(float x, float y, float z, int layer)
+bool dxfExport::point(double x, double y, double z, int layer)
 {
 	m_file << "  " << 0 << std::endl; // entity type
 	m_file << "POINT" << std::endl;
@@ -117,7 +117,7 @@ bool dxfExport::point(float x, float y, float z, int layer)
 	return (m_file.bad());
 }
 
-bool dxfExport::drawFace(const glm::vec3& p1, const glm::vec3& p2, const glm::vec3& p3, int layer)
+bool dxfExport::drawFace(const glm::dvec3& p1, const glm::dvec3& p2, const glm::dvec3& p3, int layer)
 {
 	m_file << "  " << 0 << std::endl; // entity type
 	m_file << "3DFACE" << std::endl;
@@ -155,7 +155,7 @@ bool dxfExport::drawFace(const glm::vec3& p1, const glm::vec3& p2, const glm::ve
 	return (m_file.bad());
 }
 
-bool dxfExport::drawLine(const glm::vec3& p1, const glm::vec3& p2, int layer)
+bool dxfExport::drawLine(const glm::dvec3& p1, const glm::dvec3& p2, int layer)
 {
 	m_file << "  " << 0 << std::endl; // entity type
 	m_file << "LINE" << std::endl;
@@ -184,10 +184,10 @@ bool dxfExport::drawLine(const glm::vec3& p1, const glm::vec3& p2, int layer)
 	return (m_file.bad());
 }
 
-bool dxfExport::drawTarget(const glm::vec3& center, int layer)
+bool dxfExport::drawTarget(const glm::dvec3& center, int layer)
 {
-	drawLine(center - glm::vec3(DELTA_CM, 0, 0), center + glm::vec3(DELTA_CM, 0, 0), layer);
-	drawLine(center - glm::vec3(0, DELTA_CM, 0), center + glm::vec3(0, DELTA_CM, 0), layer);
-	drawLine(center - glm::vec3(0, 0, DELTA_CM), center + glm::vec3(0, 0, DELTA_CM), layer);
+	drawLine(center - glm::dvec3(DELTA_CM, 0.0, 0.0), center + glm::dvec3(DELTA_CM, 0.0, 0.0), layer);
+	drawLine(center - glm::dvec3(0.0, DELTA_CM, 0.0), center + glm::dvec3(0.0, DELTA_CM, 0.0), layer);
+	drawLine(center - glm::dvec3(0.0, 0.0, DELTA_CM), center + glm::dvec3(0.0, 0.0, DELTA_CM), layer);
 	return (m_file.bad());
 }
